Add registerAndlogin component and add handleClick for register button in navbar
A register and login popup ui is shown when register button in navbar is clicked. Popup include tab navigation to switch between register form and login form.

Use tabs component of material for navigation in registerAndLogin component, and use popover component of material for handleClick of register button on navbar component.

Now, we can click on register button on navbar to show register/login form.